requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
transformers
torch
pdfplumber
openai
python-dotenv
